// src/api/types/MediaGenerationResponse.ts
// This file was auto-generated by Fern from our API Definition.

import type * as ElevenLabs from "../index.js";

export type MediaGenerationResponse =
    | ElevenLabs.MediaGenerationResponse.Completed
    | ElevenLabs.MediaGenerationResponse.Failed
    | ElevenLabs.MediaGenerationResponse.Generating
    | ElevenLabs.MediaGenerationResponse.Pending;

export namespace MediaGenerationResponse {
    export interface Completed extends ElevenLabs.MediaGenerationCompletedResponse {
        status: "completed";
    }

    export interface Failed extends ElevenLabs.MediaGenerationFailedResponse {
        status: "failed";
    }

    export interface Generating extends ElevenLabs.MediaGenerationInProgressResponse {
        status: "generating";
    }

    export interface Pending extends ElevenLabs.MediaGenerationInProgressResponse {
        status: "pending";
    }
}